fix crash
