+++ packages/kit/src/background/instance/backgroundApiInit.ext-ui-passkey.ts
@@ -1,0 +1,6 @@
+function backgroundApiInit() {
+  console.log('backgroundApiInit is empty in ext-passkey');
+  return null;
+}
+
+export default backgroundApiInit;
